+++ src/features/advanced/index.ts
@@ -13,6 +13,7 @@ import './cxpc-default-1y';
 import './cxpo-shorten-fields';
 import './finla-hide-ecd';
 import './flt-flight-status-icons';
+import './flt-flight-status-detailed';
 import './flt-hide-transponder';
 import './flt-shorten-addresses';
 import './flt-shorten-cargo-capacity';
